Printing object from CDFmodel class by simple print methods

An object from CDFmodel classes is yielded by function fitCDF. This objects carries the information of a fitted non-linear model.

An object from CDFmodelList classes is yielded by function fitCDF. This objects carries the information of one or more fitted non-linear models.

# S3 method for CDFmodel
print(x, digits = getOption("digits"), ...)

# S3 method for CDFmodelList
print(x, digits = getOption("digits"), ...)

Arguments

x

Object from class CDFmodel.

digits

Number of significant digits to be used.

Details

The definition of these class makes less complex the downstream analyses.

The definition of these class makes less complex the downstream analyses.